feat/project_details_page-bug_fix
